FREEI-3350: set /AMPUSER/ext/accountcode to each extension's userId
This will be "u" + the userId from userman.  This is required by the
realtimeapi conference app.  It is currently blank for most cases, so it
feels like a safe change.  In the few cases that it does get used by
something else, don't change it.  Those will likely be extensions that
will not be using the rtapi conference app.

Conflicts:
	Userman.class.php
	install.php