FREEPBX-18626: Fix override for objectClassPerson
adldap2 matches objects to predefined models with traits via the
objectClass attribute. When defining a custom schema with a different
objectClass for user objects the method for providing this objectClass
has to be overridden. adldap2 uses the method objectClassPerson to
determine the objectClass for users. objectClassUser is not used by the
library.

Signed-off-by: Adrian Nöthlich <adrian.noethlich@hadiko.de>